cleaned up unit tests / fixed question routes / unpopulate models on update
