+++ src/server/node/router/attachments.js
@@ -29,6 +29,9 @@ router.get('/',(req,res) => {
     if (_.has(req.query,'tag'))
         options.tags = req.query.tag;
 
+    if (_.has(req.query,'text'))
+        options.text = req.query.text;
+
     Attachment.list(options).then( (docs) => {
         if (_.isEmpty(docs))
             return Promise.resolve(docs);
